# feat: optimize STT performance with AWS Transcribe best practices