--- a/1.code/deep_vibe_node_server/public/index.html
+++ b/1.code/deep_vibe_node_server/public/index.html
@@ -358,7 +358,11 @@
                     <button id="recordBtn" class="btn btn-success" onclick="toggleRecording()">
                         🎤 음성인식 시작
                     </button>
+                    <button id="fileRecordBtn" class="btn btn-warning" onclick="toggleFileRecording()">
+                        📁 파일 녹음 시작
+                    </button>
                     <div id="transcribeStatus" style="margin-top: 10px;"></div>
+                    <div id="fileRecordStatus" style="margin-top: 10px;"></div>
                     
                     <div class="visualizer-container">
                         <canvas id="audioVisualizer"></canvas>