## 개요
AWS Transcribe 모범사례를 적용하여 실시간 STT 성능을 개선합니다.

## 주요 변경사항
- **청크 크기 최적화**: 4096 샘플(256ms) → 1024 샘플(64ms)로 변경
- **PCM 바이트 검증**: 단일 채널 PCM 데이터 무결성 확인 추가
- **실시간성 향상**: 더 작은 청크로 지연시간 감소
- **AWS 권장사항 준수**: 50-200ms 청크 크기 범위 내 최적화

## 기술적 개선사항
### Frontend (app.js)
- `createScriptProcessor` 버퍼 크기를 1024로 변경 (2의 거듭제곱 요구사항 충족)
- PCM 바이트 검증 로직 추가
- 청크 지속시간 로깅 개선

### Backend (transcribe-service.ts)
- 서버 측 PCM 청크 크기 검증 추가
- 짝수 바이트 확인으로 데이터 무결성 보장

## 성능 개선 효과
- **지연시간 감소**: 256ms → 64ms (75% 개선)
- **실시간성 향상**: 더 빠른 음성 인식 응답
- **데이터 품질**: PCM 포맷 검증으로 안정성 증대

## 테스트
- [x] 브라우저에서 마이크 권한 정상 동작
- [x] ScriptProcessorNode 버퍼 크기 오류 해결
- [x] AWS Transcribe 연동 정상 동작